requirements.txt
streamlit
PyPDF2
langchain_text_splitters
google-generativeai
dotenv
